refactor: add ComputerRepository
