+++ Program.cs
@@ -1,9 +1,12 @@
 ﻿using LabManager.Database;
+using LabManager.Repositories;
 using Microsoft.Data.Sqlite;
 
-
+    
 var databaseConfig = new DatabaseConfig();
 var databaseSetup = new DatabaseSetup(databaseConfig);
+
+var computerRepository = new ComputerRepository(databaseConfig);
 
 //routing
 var modelName = args[0];
@@ -13,22 +16,10 @@ if(modelName == "Computer")
 {
     if(modelAction == "List")
     {
-        var connection =  new SqliteConnection("Data Source=database_test.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers";
-
-        var reader = command.ExecuteReader();
-
-        Console.WriteLine("Computer List");
-        while(reader.Read())
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine(
-                "{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2));
+            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
         }
-
-        reader.Close();
-        connection.Close();
     }
 }
